qt: generalize wizard HWW xpub
