+++ electrum/plugins/trezor/qt.py
@@ -17,7 +17,7 @@ from electrum.plugins.hw_wallet.plugin import only_hook_if_libraries_available
 
 from electrum.gui.qt.util import (WindowModalDialog, WWLabel, Buttons, CancelButton,
                                   OkButton, CloseButton, PasswordLineEdit, getOpenFileName, ChoiceWidget)
-from electrum.gui.qt.wizard.wallet import WCScriptAndDerivation, WCHWUnlock
+from electrum.gui.qt.wizard.wallet import WCScriptAndDerivation, WCHWUnlock, WCHWXPub
 from electrum.gui.qt.wizard.wizard import WizardComponent
 
 from .trezor import (TrezorPlugin, TIM_NEW, TIM_RECOVER, TrezorInitSettings,
@@ -473,11 +473,11 @@ class Plugin(TrezorPlugin, QtPlugin):
     def extend_wizard(self, wizard: 'QENewWalletWizard'):
         super().extend_wizard(wizard)
         views = {
-            'trezor_start': { 'gui': WCScriptAndDerivation },
-            'trezor_xpub': { 'gui': WCTrezorXPub },
-            'trezor_not_initialized': { 'gui': WCTrezorInitMethod },
-            'trezor_choose_new_recover': { 'gui': WCTrezorInitParams },
-            'trezor_do_init': { 'gui': WCTrezorInit },
+            'trezor_start': {'gui': WCScriptAndDerivation},
+            'trezor_xpub': {'gui': WCTrezorXPub},
+            'trezor_not_initialized': {'gui': WCTrezorInitMethod},
+            'trezor_choose_new_recover': {'gui': WCTrezorInitParams},
+            'trezor_do_init': {'gui': WCTrezorInit},
             'trezor_unlock': {'gui': WCHWUnlock},
         }
         wizard.navmap_merge(views)
@@ -801,73 +801,12 @@ class SettingsDialog(WindowModalDialog):
         invoke_client(None)
 
 
-class WCTrezorXPub(WizardComponent, Logger):
+class WCTrezorXPub(WCHWXPub):
     def __init__(self, parent, wizard):
-        WizardComponent.__init__(self, parent, wizard, title=_('Hardware wallet information'))
-        Logger.__init__(self)
-        self.plugins = wizard.plugins
-        self.plugin = self.plugins.get_plugin('trezor')
-        self._busy = True
+        WCHWXPub.__init__(self, parent, wizard)
 
-        self.xpub = None
-        self.root_fingerprint = None
-        self.label = None
-        self.soft_device_id = None
-
-        self.ok_l = WWLabel(_('Hardware keystore added to wallet'))
-        self.ok_l.setAlignment(Qt.AlignCenter)
-        self.layout().addWidget(self.ok_l)
-
-    def on_ready(self):
-        _name, _info = self.wizard_data['hardware_device']
-        device_id = _info.device.id_
-        client = self.plugins.device_manager.client_by_id(device_id, scan_now=False)
-        client.handler = self.plugin.create_handler(self.wizard)
-
-        cosigner = self.wizard.current_cosigner(self.wizard_data)
-        xtype = cosigner['script_type']
-        derivation = cosigner['derivation_path']
-
-        def get_xpub_task(client, derivation, xtype):
-            try:
-                self.xpub = client.get_xpub(derivation, xtype, True)
-                self.root_fingerprint = client.request_root_fingerprint_from_device()
-                self.label = client.label()
-                self.soft_device_id = client.get_soft_device_id()
-            except Exception as e:
-                # TODO: handle user interaction exceptions (e.g. invalid pin) more gracefully
-                self.error = repr(e)
-                self.logger.error(repr(e))
-            self.xpub_done()
-
-        t = threading.Thread(target=get_xpub_task, args=(client, derivation, xtype), daemon=True)
-        t.start()
-
-    def xpub_done(self):
-        self.logger.debug(f'Done retrieve xpub: {self.xpub}')
-        self.busy = False
-        self.validate()
-
-    def validate(self):
-        if self.xpub and not self.error:
-            self.apply()
-            valid, error = self.wizard.check_multisig_constraints(self.wizard_data)
-            if not valid:
-                self.error = '\n'.join([
-                    _('Could not add hardware keystore to wallet'),
-                    error
-                ])
-            self.valid = valid
-        else:
-            self.valid = False
-
-    def apply(self):
-        cosigner_data = self.wizard.current_cosigner(self.wizard_data)
-        cosigner_data['hw_type'] = 'trezor'
-        cosigner_data['master_key'] = self.xpub
-        cosigner_data['root_fingerprint'] = self.root_fingerprint
-        cosigner_data['label'] = self.label
-        cosigner_data['soft_device_id'] = self.soft_device_id
+    def get_xpub_from_client(self, client, derivation, xtype):
+        return client.get_xpub(derivation, xtype, True)
 
 
 class WCTrezorInitMethod(WizardComponent, Logger):
